photo deleting

+++ src/main/java/com/example/communication/controller/MainPageController.java
@@ -1,8 +1,10 @@
 package com.example.communication.controller;
 
 import com.example.communication.model.Message;
+import com.example.communication.model.Role;
 import com.example.communication.model.User;
 import com.example.communication.model.dto.MessageDTO;
+import com.example.communication.repository.MessageRepository;
 import com.example.communication.service.MessageService;
 import java.io.IOException;
 import java.util.List;
@@ -25,8 +27,11 @@ public class MainPageController {
 
   private final MessageService messageService;
 
-  public MainPageController(MessageService messageService) {
+  private final MessageRepository messageRepository;
+
+  public MainPageController(MessageService messageService, MessageRepository messageRepository) {
     this.messageService = messageService;
+    this.messageRepository = messageRepository;
   }
 
   @GetMapping("/users")
@@ -41,6 +46,8 @@ public class MainPageController {
       Model model
   ) {
     List<MessageDTO> messages = messageService.getAllMessages(filter, user);
+    model.addAttribute("loginUserId", user.getId());
+    model.addAttribute("isAdmin", user.getRoles().contains(Role.ADMIN));
     model.addAttribute("messages", messages);
     model.addAttribute("filter", "");
     return "main";
@@ -62,8 +69,11 @@ public class MainPageController {
 
   @GetMapping("/delete/{id}")
   public String delete(
+      @AuthenticationPrincipal User user,
       @PathVariable(value="id") Long id
   ) {
+    if (!user.getRoles().contains(Role.ADMIN) && !user.getId().equals(messageRepository.getOne(id).getUser().getId()))
+      return "redirect:/";
     ControllerUtils.deleteMessage(id);
     return "redirect:/";
   }
